*** empty log message ***

+++ coders/8bim.c
@@ -76,8 +76,7 @@ static unsigned int
 %
 %  The format of the Is8BIM method is:
 %
-%      unsigned int Is8BIM(const unsigned char *magick,
-%        const unsigned int length)
+%      unsigned int Is8BIM(const unsigned char *magick,const size_t length)
 %
 %  A description of each parameter follows:
 %
@@ -90,8 +89,7 @@ static unsigned int
 %
 %
 */
-static unsigned int Is8BIM(const unsigned char *magick,
-  const unsigned int length)
+static unsigned int Is8BIM(const unsigned char *magick,const size_t length)
 {
   if (length < 4)
     return(False);
