+++ coders/bmp.c
@@ -354,8 +354,7 @@ static unsigned int EncodeImage(Image *image,const unsigned int bytes_per_line,
 %
 %  The format of the IsBMP method is:
 %
-%      unsigned int IsBMP(const unsigned char *magick,
-%        const unsigned int length)
+%      unsigned int IsBMP(const unsigned char *magick,const size_t length)
 %
 %  A description of each parameter follows:
 %
@@ -368,7 +367,7 @@ static unsigned int EncodeImage(Image *image,const unsigned int bytes_per_line,
 %
 %
 */
-static unsigned int IsBMP(const unsigned char *magick,const unsigned int length)
+static unsigned int IsBMP(const unsigned char *magick,const size_t length)
 {
   if (length < 2)
     return(False);
