+++ coders/dcm.c
@@ -2616,8 +2616,7 @@ static const DicomInfo
 %
 %  The format of the ReadDCMImage method is:
 %
-%      unsigned int IsDCM(const unsigned char *magick,
-%        const unsigned int length)
+%      unsigned int IsDCM(const unsigned char *magick,const size_t length)
 %
 %  A description of each parameter follows:
 %
@@ -2630,7 +2629,7 @@ static const DicomInfo
 %
 %
 */
-static unsigned int IsDCM(const unsigned char *magick,const unsigned int length)
+static unsigned int IsDCM(const unsigned char *magick,const size_t length)
 {
   if (length < 132)
     return(False);
