+++ coders/hdf.c
@@ -80,8 +80,7 @@ static unsigned int
 %
 %  The format of the IsHDF method is:
 %
-%      unsigned int IsHDF(const unsigned char *magick,
-%        const unsigned int length)
+%      unsigned int IsHDF(const unsigned char *magick,const size_t length)
 %
 %  A description of each parameter follows:
 %
@@ -94,7 +93,7 @@ static unsigned int
 %
 %
 */
-static unsigned int IsHDF(const unsigned char *magick,const unsigned int length)
+static unsigned int IsHDF(const unsigned char *magick,const size_t length)
 {
   if (length < 4)
     return(False);
