+++ coders/jp2.c
@@ -83,8 +83,7 @@ static unsigned int
 %
 %  The format of the IsJP2 method is:
 %
-%      unsigned int IsJP2(const unsigned char *magick,
-%        const unsigned int length)
+%      unsigned int IsJP2(const unsigned char *magick,const size_t length)
 %
 %  A description of each parameter follows:
 %
@@ -97,7 +96,7 @@ static unsigned int
 %
 %
 */
-static unsigned int IsJP2(const unsigned char *magick,const unsigned int length)
+static unsigned int IsJP2(const unsigned char *magick,const size_t length)
 {
   if (length < 9)
     return(False);
@@ -123,7 +122,7 @@ static unsigned int IsJP2(const unsigned char *magick,const unsigned int length)
 %  The format of the IsJPC method is:
 %
 %      unsigned int IsJPC(const unsigned char *magick,
-%        const unsigned int length)
+%        const size_t length)
 %
 %  A description of each parameter follows:
 %
@@ -136,7 +135,7 @@ static unsigned int IsJP2(const unsigned char *magick,const unsigned int length)
 %
 %
 */
-static unsigned int IsJPC(const unsigned char *magick,const unsigned int length)
+static unsigned int IsJPC(const unsigned char *magick,const size_t length)
 {
   if (length < 2)
     return(False);
