+++ coders/miff.c
@@ -83,8 +83,7 @@ static unsigned int
 %
 %  The format of the IsMIFF method is:
 %
-%      unsigned int IsMIFF(const unsigned char *magick,
-%        const unsigned int length)
+%      unsigned int IsMIFF(const unsigned char *magick,const size_t length)
 %
 %  A description of each parameter follows:
 %
@@ -97,8 +96,7 @@ static unsigned int
 %
 %
 */
-static unsigned int IsMIFF(const unsigned char *magick,
-  const unsigned int length)
+static unsigned int IsMIFF(const unsigned char *magick,const size_t length)
 {
   if (length < 14)
     return(False);
