+++ coders/mpc.c
@@ -78,7 +78,7 @@ static unsigned int
 %
 %  The format of the IsMPC method is:
 %
-%      unsigned int IsMPC(const unsigned char *magick,const unsigned int length)
+%      unsigned int IsMPC(const unsigned char *magick,const size_t length)
 %
 %  A description of each parameter follows:
 %
@@ -91,7 +91,7 @@ static unsigned int
 %
 %
 */
-static unsigned int IsMPC(const unsigned char *magick,const unsigned int length)
+static unsigned int IsMPC(const unsigned char *magick,const size_t length)
 {
   if (length < 14)
     return(False);
