+++ coders/pdf.c
@@ -82,8 +82,7 @@ static unsigned int
 %
 %  The format of the IsPDF method is:
 %
-%      unsigned int IsPDF(const unsigned char *magick,
-%        const unsigned int length)
+%      unsigned int IsPDF(const unsigned char *magick,const size_t length)
 %
 %  A description of each parameter follows:
 %
@@ -96,7 +95,7 @@ static unsigned int
 %
 %
 */
-static unsigned int IsPDF(const unsigned char *magick,const unsigned int length)
+static unsigned int IsPDF(const unsigned char *magick,const size_t length)
 {
   if (length < 5)
     return(False);
