+++ coders/sct.c
@@ -71,8 +71,7 @@
 %
 %  The format of the IsSCT method is:
 %
-%      unsigned int IsSCT(const unsigned char *magick,
-%        const unsigned int length)
+%      unsigned int IsSCT(const unsigned char *magick,const size_t length)
 %
 %  A description of each parameter follows:
 %
@@ -85,7 +84,7 @@
 %
 %
 */
-static unsigned int IsSCT(const unsigned char *magick,const unsigned int length)
+static unsigned int IsSCT(const unsigned char *magick,const size_t length)
 {
   if (length < 2)
     return(False);
