+++ coders/sgi.c
@@ -102,8 +102,7 @@ static unsigned int
 %
 %  The format of the IsSGI method is:
 %
-%      unsigned int IsSGI(const unsigned char *magick,
-%        const unsigned int length)
+%      unsigned int IsSGI(const unsigned char *magick,const size_t length)
 %
 %  A description of each parameter follows:
 %
@@ -116,7 +115,7 @@ static unsigned int
 %
 %
 */
-static unsigned int IsSGI(const unsigned char *magick,const unsigned int length)
+static unsigned int IsSGI(const unsigned char *magick,const size_t length)
 {
   if (length < 2)
     return(False);
