+++ coders/svg.c
@@ -181,8 +181,7 @@ static unsigned int
 %
 %  The format of the IsSVG method is:
 %
-%      unsigned int IsSVG(const unsigned char *magick,
-%        const unsigned int length)
+%      unsigned int IsSVG(const unsigned char *magick,const size_t length)
 %
 %  A description of each parameter follows:
 %
@@ -195,8 +194,7 @@ static unsigned int
 %
 %
 */
-static unsigned int IsSVG(const unsigned char *magick,
-  const unsigned int length)
+static unsigned int IsSVG(const unsigned char *magick,const size_t length)
 {
   if (length < 5)
     return(False);
