+++ coders/tiff.c
@@ -92,8 +92,7 @@ static unsigned int
 %
 %  The format of the IsTIFF method is:
 %
-%      unsigned int IsTIFF(const unsigned char *magick,
-%        const unsigned int length)
+%      unsigned int IsTIFF(const unsigned char *magick,const size_t length)
 %
 %  A description of each parameter follows:
 %
@@ -106,8 +105,7 @@ static unsigned int
 %
 %
 */
-static unsigned int IsTIFF(const unsigned char *magick,
-  const unsigned int length)
+static unsigned int IsTIFF(const unsigned char *magick,const size_t length)
 {
   if (length < 4)
     return(False);
