+++ coders/xbm.c
@@ -77,8 +77,7 @@ static unsigned int
 %
 %  The format of the IsXBM method is:
 %
-%      unsigned int IsXBM(const unsigned char *magick,
-%        const unsigned int length)
+%      unsigned int IsXBM(const unsigned char *magick,const size_t length)
 %
 %  A description of each parameter follows:
 %
@@ -91,7 +90,7 @@ static unsigned int
 %
 %
 */
-static unsigned int IsXBM(const unsigned char *magick,const unsigned int length)
+static unsigned int IsXBM(const unsigned char *magick,const size_t length)
 {
   if (length < 7)
     return(False);
